Tables: move table Column classes into their own sub-namespace
