+++ src/Tables/Columns/Column.php
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-namespace Gibbon\Tables;
+namespace Gibbon\Tables\Columns;
 
 use Gibbon\Forms\Traits\BasicAttributesTrait;
