+++ src/Tables/DataTable.php
@@ -22,8 +22,10 @@ namespace Gibbon\Tables;
 use Gibbon\Domain\DataSet;
 use Gibbon\Domain\QueryCriteria;
 use Gibbon\Tables\Action;
-use Gibbon\Tables\Column;
-use Gibbon\Tables\ActionColumn;
+use Gibbon\Tables\Columns\Column;
+use Gibbon\Tables\Columns\ActionColumn;
+use Gibbon\Tables\Columns\CheckboxColumn;
+use Gibbon\Tables\Columns\ExpandableColumn;
 use Gibbon\Tables\Renderer\SimpleRenderer;
 use Gibbon\Tables\Renderer\PaginatedRenderer;
 use Gibbon\Tables\Renderer\RendererInterface;
